+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 * Read/Write (rw) and read-only (r) access have been added to api keys - Update apikeys.json
 * Added /api/[index]/flush to the api.
 * Added count query to socket client.
+* Fixed default analyzer.
+* Added support for default indexer